fix: unwrap or default
Recently I ran into multiple panics due to unwrap not being able to
return a value.

In order to fix the exception=PanicException('called `Option::unwrap()`
on a `None` value')>, all unwraps have been transformed to either unwrap
or default to return either None or the value itself, which is
considered a best practice.

EDIT: Added anyhow to propagate errors to upstream

Signed-off-by: Sebastian Waldbauer <sebastian.waldbauer@posedio.com>